Crate reth_primitives

Commonly used types in reth.

This crate contains Ethereum primitive types and helper functions.

Feature Flags

arbitrary: Adds proptest and arbitrary support for primitive types.

test-utils: Export utilities for testing

Re-exports

pub use constants::DEV_GENESIS_HASH;

pub use constants::EMPTY_OMMER_ROOT_HASH;

pub use constants::GOERLI_GENESIS_HASH;

pub use constants::HOLESKY_GENESIS_HASH;

pub use constants::KECCAK_EMPTY;

pub use constants::MAINNET_GENESIS_HASH;

pub use constants::SEPOLIA_GENESIS_HASH;

pub use snapshot::SnapshotSegment;

pub use transaction::BlobTransaction;

pub use transaction::BlobTransactionSidecar;

pub use transaction::BlobTransactionValidationError;

pub use transaction::FromRecoveredPooledTransaction;

pub use transaction::PooledTransactionsElement;

pub use transaction::PooledTransactionsElementEcRecovered;

pub use transaction::AccessList;

pub use transaction::AccessListItem;

pub use transaction::FromRecoveredTransaction;

pub use transaction::IntoRecoveredTransaction;

pub use transaction::InvalidTransactionError;

pub use transaction::Signature;

pub use transaction::Transaction;

pub use transaction::TransactionKind;

pub use transaction::TransactionMeta;

pub use transaction::TransactionSigned;

pub use transaction::TransactionSignedEcRecovered;

pub use transaction::TransactionSignedNoHash;

pub use transaction::TxEip1559;

pub use transaction::TxEip2930;

pub use transaction::TxEip4844;

pub use transaction::TxHashOrNumber;

pub use transaction::TxLegacy;

pub use transaction::TxType;

pub use transaction::TxValue;

pub use transaction::EIP1559_TX_TYPE_ID;

pub use transaction::EIP2930_TX_TYPE_ID;

pub use transaction::EIP4844_TX_TYPE_ID;

pub use transaction::LEGACY_TX_TYPE_ID;

pub use alloy_primitives;

pub use alloy_primitives::bytes;

pub use alloy_primitives::hex;

pub use alloy_primitives::hex_literal;

pub use alloy_primitives::ruint;

pub use revm_primitives;

pub use arbitrary;

pub use c_kzg as kzg;

Modules

arbitrary

The Arbitrary trait crate.

basefee

Helpers for working with EIP-1559 base fee

constants

Ethereum protocol-related constants

eip4844

Helpers for working with EIP-4844 blob fee.

fs

Wrapper for std::fs methods

genesis

Types for genesis configuration of a chain.

proofs

Helper function for calculating Merkle proofs and hashes.

revm

Helpers for working with revm

serde_helper

serde utilities.

snapshot

Snapshot primitives.

stage

Staged sync primitives.

transaction

Helpers for working with transactions

trie

Collection of trie related types.

Macros

address

Converts a sequence of string literals containing hex-encoded data into a new Address at compile time.

b256

Converts a sequence of string literals containing hex-encoded data into a new B256 at compile time.

bloom

Converts a sequence of string literals containing hex-encoded data into a new Bloom at compile time.

bytes

Converts a sequence of string literals containing hex-encoded data into a new Bytes at compile time.

hex

Macro for converting sequence of string literals containing hex-encoded data into an array of bytes.

Structs

Account

An Ethereum account.

Address

An Ethereum address, 20 bytes in length.

BaseFeeParams

BaseFeeParams contains the config parameters that control block base fee computation

Block

Ethereum full block.

BlockBody

A response to GetBlockBodies, containing bodies if any bodies were found.

BlockNumHash

Block number and hash.

BlockWithSenders

Sealed block with senders recovered from transactions.

Bloom

Ethereum 256 byte bloom filter.

Bytecode

Bytecode for an account.

Bytes

Wrapper type around bytes::Bytes to support “0x” prefixed hex strings.

BytesMut

A unique reference to a contiguous slice of memory.

Chain

Either a known NamedChain or a EIP-155 chain ID.

ChainConfig

Defines core blockchain settings per block.

ChainInfo

Current status of the blockchain’s head.

ChainSpec

An Ethereum chain specification.

ChainSpecBuilder

A helper to build custom chain specs

DisplayHardforks

A container for pretty-printing a list of hardforks.

ForkBaseFeeParams

A type alias to a vector of tuples of Hardfork and BaseFeeParams, sorted by Hardfork activation order. This is used to specify dynamic EIP-1559 parameters for chains like Optimism.

ForkFilter

Filter that describes the state of blockchain and can be used to check incoming ForkIds for compatibility.

ForkHash

CRC32 hash of all previous forks starting from genesis block.

ForkId

A fork identifier as defined by EIP-2124. Serves as the chain compatibility identifier.

ForkTimestamps

Various timestamps of forks

ForkTransition

Represents a transition from one fork to another

Genesis

The genesis block specification.

GenesisAccount

An account in the state of the genesis block.

GotExpected

A pair of values, one of which is expected and one of which is actual.

GotExpectedBoxed

A pair of values, one of which is expected and one of which is actual.

Head

Describes the current head block.

Header

Block header

IntegerList

Uses EliasFano to hold a list of integers. It provides really good compression with the capability to access its elements without decoding it.

JumpMap

A map of valid jump destinations.

Log

Ethereum Log

NodeRecord

Represents a ENR in discovery.

PruneCheckpoint

Saves the pruning progress of a stage.

PruneModes

Pruning configuration for every segment of the data that can be pruned.

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

ReceiptWithBloomRef

Receipt reference type with calculated bloom filter.

Receipts

A collection of receipts organized as a two-dimensional vector.

ReceiptsLogPruneConfig

Configuration for pruning receipts not associated with logs emitted by the specified contracts.

ReusableDecompressor

Reusable decompressor that uses its own internal buffer.

RpcBlockHash

A block hash which may have a boolean requireCanonical field. If false, an RPC call should raise if a block matching the hash is not found. If true, an RPC call should additionaly raise if the block is not in the canonical chain. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md#specification

SealedBlock

Sealed Ethereum full block.

SealedBlockWithSenders

Sealed block with senders recovered from transactions.

SealedHeader

A Header that is sealed at a precalculated hash, use SealedHeader::unseal() if you want to modify header.

StorageEntry

Account storage entry.

WithPeerId

Generic wrapper with peer id

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Withdrawals

Represents a collection of Withdrawals.

Enums

AllGenesisFormats

A helper type for compatibility with geth’s config

BaseFeeParamsKind

A wrapper around BaseFeeParams that allows for specifying constant or dynamic EIP-1559 parameters based on the active Hardfork.

BlockHashOrNumber

Either a block hash or a block number

BlockId

A Block Identifier https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

BloomInput

Input to the Bloom::accrue method.

ForkCondition

The condition at which a fork is activated.

ForkFilterKey

How to filter forks.

Hardfork

The name of an Ethereum hardfork.

HeadersDirection

Represents the direction for a headers request depending on the reverse field of the request.

NamedChain

An Ethereum EIP-155 chain.

PruneMode

Prune mode.

PruneProgress

Progress of pruning.

PruneSegment

Segment of the data that can be pruned.

PruneSegmentError

PruneSegment error type.

ValidationError

Reason for rejecting provided ForkId.

Constants

MINIMUM_PRUNING_DISTANCE

Minimum distance from the tip necessary for the node to work correctly:

RECEIPT_COMPRESSOR

Thread receipt compressor.

RECEIPT_DECOMPRESSOR

Thread receipt decompressor.

TRANSACTION_COMPRESSOR

Thread Transaction compressor.

TRANSACTION_DECOMPRESSOR

Thread Transaction decompressor.

Statics

DEV

Dev testnet specification

GOERLI

The Goerli spec

GOERLI_BOOTNODES

GOERLI bootnodes

HOLESKY

The Holesky spec

HOLESKY_BOOTNODES

Ethereum Foundation Holesky BOOTNODES

MAINNET

The Ethereum mainnet spec

MAINNET_BOOTNODES

Ethereum Foundation Go Bootnodes

RECEIPT_DICTIONARY

Compression/Decompression dictionary for Receipt.

SEPOLIA

The Sepolia spec

SEPOLIA_BOOTNODES

Ethereum Foundation SEPOLIA BOOTNODES

TRANSACTION_DICTIONARY

Compression/Decompression dictionary for Transaction.

Traits

Buf

Read bytes from a buffer.

BufMut

A trait for values that provide sequential write access to bytes.

UintTryTo

⚠️ Workaround for Rust issue #50133. Use TryFrom instead.

Functions

eip191_hash_message

Hash a message according to EIP-191 (version 0x01).

goerli_nodes

Returns parsed goerli nodes

holesky_nodes

Returns parsed holesky nodes

keccak256

Simple interface to the Keccak-256 hash function.

logs_bloom

Calculate receipt logs bloom.

mainnet_nodes

Returns parsed mainnet nodes

parse_nodes

Parses all the nodes

public_key_to_address

Converts a public key into an ethereum address by hashing the encoded public key with keccak256.

recover_signer_unchecked

Recovers the address of the sender using secp256k1 pubkey recovery.

sepolia_nodes

Returns parsed sepolia nodes

sign_message

Signs message with the given secret key. Returns the corresponding signature.

Type Aliases

B64

8-byte fixed byte-array type.

B128

16-byte fixed byte-array type.

B256

32-byte fixed byte-array type.

B512

64-byte fixed byte-array type.

BlockHash

A block hash.

BlockNumber

A block number.

ChainId

Chain identifier type (introduced in EIP-155).

ForkBlock

Block number and hash of the forked block.

PeerId

Alias for a peer identifier

Selector

Solidity contract functions are addressed using the first four bytes of the Keccak-256 hash of their signature.

StorageKey

An account storage key.

StorageValue

An account storage value.

TxHash

A transaction hash is a keccak hash of an RLP encoded signed transaction.

TxIndex

The index of transaction in a block.

TxNumber

The sequence number of all existing transactions.

U8

Uint for 8 bits. Similar to u8.

U64

Uint for 64 bits. Similar to u64.

U128

Uint for 128 bits. Similar to u128.

U256

Uint for 256 bits.