Module reth_primitives::constants

Ethereum protocol-related constants

Modules

eip4844

EIP-4844 constants. EIP-4844 protocol constants and utils for shard Blob Transactions.

Constants

ALLOWED_FUTURE_BLOCK_TIME_SECONDS

Max seconds from current time allowed for blocks, before they’re considered future blocks.

BEACON_CONSENSUS_REORG_UNWIND_DEPTH

The number of blocks to unwind during a reorg that already became a part of canonical chain.

BEACON_NONCE

The default block nonce in the beacon consensus

BEACON_ROOTS_ADDRESS

The address for the beacon roots contract defined in EIP-4788.

DEV_GENESIS_HASH

Testnet genesis hash.

EIP1559_DEFAULT_BASE_FEE_MAX_CHANGE_DENOMINATOR

Base fee max change denominator as defined in EIP-1559

EIP1559_DEFAULT_ELASTICITY_MULTIPLIER

Elasticity multiplier as defined in EIP-1559

EIP1559_INITIAL_BASE_FEE

Initial base fee as defined in EIP-1559

EMPTY_OMMER_ROOT_HASH

Ommer root of empty list.

EMPTY_RECEIPTS

Transactions root of empty receipts set.

EMPTY_ROOT_HASH

Root hash of an empty trie.

EMPTY_TRANSACTIONS

Transactions root of empty transactions set.

EMPTY_WITHDRAWALS

Withdrawals root of empty withdrawals set.

EPOCH_DURATION

An EPOCH is a series of 32 slots (~6.4min).

EPOCH_SLOTS

An EPOCH is a series of 32 slots.

ETHEREUM_BLOCK_GAS_LIMIT

The default Ethereum block gas limit. See https://github.com/paradigmxyz/reth/issues/3233.

ETH_TO_WEI

Multiplier for converting ether to wei.

FINNEY_TO_WEI

Multiplier for converting finney (milliether) to wei.

GOERLI_BASE_GENESIS

Base goerli genesis hash.

GOERLI_GENESIS_HASH

Goerli genesis hash.

GOERLI_OP_GENESIS

Optimism goerli genesis hash.

GWEI_TO_WEI

Multiplier for converting gwei to wei.

HOLESKY_GENESIS_HASH

Holesky genesis hash.

KECCAK_EMPTY

Keccak256 over empty array.

MAINNET_GENESIS_HASH

The Ethereum mainnet genesis hash.

MAXIMUM_EXTRA_DATA_SIZE

Maximum extra data size in a block after genesis

MGAS_TO_GAS

Multiplier for converting mgas to gas.

MIN_PROTOCOL_BASE_FEE

The minimum tx fee below which the txpool will reject the transaction.

MIN_PROTOCOL_BASE_FEE_U256

Same as MIN_PROTOCOL_BASE_FEE but as a U256.

RETH_CLIENT_VERSION

The client version: reth/v{major}.{minor}.{patch}

SELECTOR_LEN

The first four bytes of the call data for a function call specifies the function to be called.

SEPOLIA_GENESIS_HASH

Sepolia genesis hash.

SLOT_DURATION

The duration of a slot in seconds.

SYSTEM_ADDRESS

The caller to be used when calling the EIP-4788 beacon roots contract at the beginning of the block.