Macro reth_primitives::bloom

macro_rules! bloom {
    () => { ... };
    ($($s:literal)+) => { ... };
}

Converts a sequence of string literals containing hex-encoded data into a new Bloom at compile time.

If the input is empty, a zero-initialized array is returned.

Note that the strings cannot be prefixed with 0x.

See hex! for more information.

Examples

use alloy_primitives::{bloom, Bloom};

const ZERO: Bloom = bloom!();
assert_eq!(ZERO, Bloom::ZERO);

let byte_array: Bloom = bloom!("…");