Module reth_primitives::transaction

Helpers for working with transactions

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

Represents a list of addresses and storage keys that a transaction plans to access.

BlobTransaction

A response to GetPooledTransactions that includes blob data, their commitments, and their corresponding proofs.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

PooledTransactionsElementEcRecovered

A signed pooled transaction with recovered signer.

Signature

r, s: Values corresponding to the signature of the transaction and used to determine the sender of the transaction; formally Tr and Ts. This is expanded in Appendix F of yellow paper.

TransactionMeta

Additional fields in the context of a block that contains this transaction.

TransactionSigned

Signed transaction.

TransactionSignedEcRecovered

Signed transaction with recovered signer.

TransactionSignedNoHash

Signed transaction without its Hash. Used type for inserting into the DB.

TxEip1559

A transaction with a priority fee (EIP-1559).

TxEip2930

Transaction with an AccessList (EIP-2930).

TxEip4844

EIP-4844 Blob Transaction

TxLegacy

Legacy transaction.

TxValue

TxValue is the type of the value field in the various Ethereum transactions structs.

Enums

BlobTransactionValidationError

An error that can occur when validating a BlobTransaction.

InvalidTransactionError

Represents error variants that can happen when trying to validate a Transaction

PooledTransactionsElement

A response to GetPooledTransactions. This can include either a blob transaction, or a non-4844 signed transaction.

Transaction

A raw transaction.

TransactionKind

Whether or not the transaction is a contract creation.

TransactionSignedVariant

Represents various different transaction formats used in reth.

TxType

Transaction Type

Constants

EIP1559_TX_TYPE_ID

Identifier for TxEip1559 transaction.

EIP2930_TX_TYPE_ID

Identifier for TxEip2930 transaction.

EIP4844_TX_TYPE_ID

Identifier for TxEip4844 transaction.

LEGACY_TX_TYPE_ID

Identifier for legacy transaction, however TxLegacy this is technically not typed.

Traits

FromRecoveredPooledTransaction

A transaction type that can be created from a PooledTransactionsElementEcRecovered transaction.

FromRecoveredTransaction

A transaction type that can be created from a TransactionSignedEcRecovered transaction.

IntoRecoveredTransaction

The inverse of FromRecoveredTransaction that ensure the transaction can be sent over the network

Type Aliases

TxHashOrNumber

Either a transaction hash or number.