Struct reth_primitives::transaction::TransactionSigned

pub struct TransactionSigned {
    pub hash: TxHash,
    pub signature: Signature,
    pub transaction: Transaction,
}

Signed transaction.

Fields
hash: TxHash

Transaction hash

signature: Signature

The transaction signature values

transaction: Transaction

Raw transaction info

Implementations

impl TransactionSigned

pub fn signature(&self) -> &Signature

Transaction signature.

pub fn hash(&self) -> TxHash

Transaction hash. Used to identify transaction.

pub fn hash_ref(&self) -> &TxHash

Reference to transaction hash. Used to identify transaction.

pub fn recover_signer(&self) -> Option<Address>

Recover signer from signature and hash.

Returns None if the transaction’s signature is invalid, see also Self::recover_signer.

pub fn recover_signer_unchecked(&self) -> Option<Address>

Recover signer from signature and hash without ensuring that the signature has a low s value.

Returns None if the transaction’s signature is invalid, see also Self::recover_signer_unchecked.

pub fn recover_signers<'a, T>(txes: T, num_txes: usize) -> Option<Vec<Address>>
where T: IntoParallelIterator<Item = &'a Self> + IntoIterator<Item = &'a Self> + Send,

Recovers a list of signers from a transaction list iterator.

Returns None, if some transaction’s signature is invalid, see also Self::recover_signer.

pub fn recover_signers_unchecked<'a, T>( txes: T, num_txes: usize ) -> Option<Vec<Address>>
where T: IntoParallelIterator<Item = &'a Self> + IntoIterator<Item = &'a Self> + Send,

Recovers a list of signers from a transaction list iterator without ensuring that the signature has a low s value.

Returns None, if some transaction’s signature is invalid, see also Self::recover_signer_unchecked.

pub const fn with_signer(self, signer: Address) -> TransactionSignedEcRecovered

Returns the TransactionSignedEcRecovered transaction with the given sender.

pub fn into_ecrecovered(self) -> Option<TransactionSignedEcRecovered>

Consumes the type, recover signer and return TransactionSignedEcRecovered

Returns None if the transaction’s signature is invalid, see also Self::recover_signer.

pub fn into_ecrecovered_unchecked(self) -> Option<TransactionSignedEcRecovered>

Consumes the type, recover signer and return TransactionSignedEcRecovered without ensuring that the signature has a low s value (EIP-2).

Returns None if the transaction’s signature is invalid, see also Self::recover_signer_unchecked.

pub fn try_ecrecovered(&self) -> Option<TransactionSignedEcRecovered>

Tries to recover signer and return TransactionSignedEcRecovered by cloning the type.

pub fn try_into_ecrecovered(self) -> Result<TransactionSignedEcRecovered, Self>

Tries to recover signer and return TransactionSignedEcRecovered.

Returns Err(Self) if the transaction’s signature is invalid, see also Self::recover_signer.

pub fn try_into_ecrecovered_unchecked( self ) -> Result<TransactionSignedEcRecovered, Self>

Tries to recover signer and return TransactionSignedEcRecovered. without ensuring that the signature has a low s value (EIP-2).

Returns Err(Self) if the transaction’s signature is invalid, see also Self::recover_signer_unchecked.

pub fn envelope_encoded(&self) -> Bytes

Returns the enveloped encoded transactions.

See also TransactionSigned::encode_enveloped

pub fn encode_enveloped(&self, out: &mut dyn BufMut)

Encodes the transaction into the “raw” format (e.g. eth_sendRawTransaction). This format is also referred to as “binary” encoding.

For legacy transactions, it encodes the RLP of the transaction into the buffer: rlp(tx-data) For EIP-2718 typed it encodes the type of the transaction followed by the rlp of the transaction: tx-type || rlp(tx-data)

pub fn recalculate_hash(&self) -> B256

Calculate transaction hash, eip2728 transaction does not contain rlp header and start with tx type.

pub fn from_transaction_and_signature( transaction: Transaction, signature: Signature ) -> Self

Create a new signed transaction from a transaction and its signature. This will also calculate the transaction hash using its encoding.

pub fn size(&self) -> usize

Calculate a heuristic for the in-memory size of the TransactionSigned.

pub fn decode_rlp_legacy_transaction( data: &mut &[u8] ) -> Result<TransactionSigned>

Decodes legacy transaction from the data buffer.

This should be used only be used in general transaction decoding methods, which have already ensured that the input is a legacy transaction with the following format: rlp(legacy_tx)

Legacy transactions are encoded as lists, so the input should start with a RLP list header.

This expects rlp(legacy_tx)

pub fn decode_enveloped_typed_transaction( data: &mut &[u8] ) -> Result<TransactionSigned>

Decodes en enveloped EIP-2718 typed transaction.

This should be used only be used internally in general transaction decoding methods, which have already ensured that the input is a typed transaction with the following format: tx-type || rlp(tx-data)

Note that this format does not start with any RLP header, and instead starts with a single byte indicating the transaction type.

CAUTION: this expects that data is tx-type || rlp(tx-data)

pub fn decode_enveloped(data: &mut &[u8]) -> Result<Self>

Decodes the “raw” format of transaction (similar to eth_sendRawTransaction).

This should be used for any RPC method that accepts a raw transaction, excluding raw EIP-4844 transactions in eth_sendRawTransaction. Currently, this includes:

eth_sendRawTransaction for non-EIP-4844 transactions.

All versions of engine_newPayload, in the transactions field.

A raw transaction is either a legacy transaction or EIP-2718 typed transaction.

For legacy transactions, the format is encoded as: rlp(tx-data). This format will start with a RLP list header.

For EIP-2718 typed transactions, the format is encoded as the type of the transaction followed by the rlp of the transaction: type || rlp(tx-data).

To decode EIP-4844 transactions from eth_sendRawTransaction, use PooledTransactionsElement::decode_enveloped.

pub fn length_without_header(&self) -> usize

Returns the length without an RLP header - this is used for eth/68 sizes.