Module reth_primitives::trie

Collection of trie related types.

Re-exports

pub use nodes::StoredBranchNode;

Modules

hash_builder

The implementation of hash builder. MPT hash builder implementation.

nodes

Various branch nodes produced by the hash builder.

Structs

AccountProof

The merkle proof with the relevant account info.

BranchNodeCompact

A struct representing a branch node in an Ethereum trie.

HashBuilder

A component used to construct the root hash of the trie. The primary purpose of a Hash Builder is to build the Merkle proof that is essential for verifying the integrity and authenticity of the trie’s contents. It achieves this by constructing the root hash from the hashes of child nodes according to specific rules, depending on the type of the node (branch, extension, or leaf).

Nibbles

Structure representing a sequence of nibbles.

StorageProof

The merkle proof of the storage entry.

StorageTrieEntry

Account storage trie node.

StoredNibbles

The representation of nibbles of the merkle trie stored in the database.

StoredNibblesSubKey

The representation of nibbles of the merkle trie stored in the database.

StoredSubNode

Walker sub node for storing intermediate state root calculation state in the database. See crate::stage::MerkleCheckpoint.

TrieAccount

An Ethereum account as represented in the trie.

TrieMask

A struct representing a mask of 16 bits, used for Ethereum trie operations.

Constants

EMPTY_ROOT_HASH

Root hash of an empty trie.