Module reth_primitives::trie::nodes

Various branch nodes produced by the hash builder.

Structs

BranchNode

A Branch node is only a pointer to the stack of nodes and is used to create the RLP encoding of the node using masks which filter from the stack of nodes.

BranchNodeCompact

A struct representing a branch node in an Ethereum trie.

ExtensionNode

An intermediate node that exists solely to compress the trie’s paths. It contains a path segment (a shared prefix of keys) and a single child pointer. Essentially, an extension node can be thought of as a shortcut within the trie to reduce its overall depth.

LeafNode

A leaf node represents the endpoint or terminal node in the trie. In other words, a leaf node is where actual values are stored.

StoredBranchNode

Wrapper around BranchNodeCompact that implements Compact.

Constants

CHILD_INDEX_RANGE

The range of valid child indexes.

Functions

word_rlp

Optimization for quick encoding of a 32-byte word as RLP.