Module reth_primitives::revm::env

The env module provides essential utilities for managing Ethereum transaction and block environments.

It includes functions to fill transaction and block environments with relevant data, handle system contract calls, and recover the signer of Ethereum headers.

Enums

CliqueSignerRecoveryError

Error type for recovering Clique signer from a header.

Functions

block_coinbase

Return the coinbase address for the given header and chain spec.

fill_block_env

Fill block environment from Block.

fill_block_env_with_coinbase

Fill block environment with coinbase.

fill_cfg_and_block_env

Convenience function to call both fill_cfg_env and fill_block_env

fill_cfg_env

Fill CfgEnv fields according to the chain spec and given header

fill_tx_env

Fill transaction environment from a Transaction and the given sender address.

fill_tx_env_with_beacon_root_contract_call

Fill transaction environment with the EIP-4788 system contract message data.

fill_tx_env_with_recovered

Fill transaction environment from TransactionSignedEcRecovered.

recover_header_signer

Recover the account from signed header per clique consensus rules.

tx_env_with_recovered

Returns a new TxEnv filled with the transaction’s data.